requirements.txt
streamlit
openai
google-api-python-client
requests
beautifulsoup4
pyautogen
